Reorganized SetAlgebra by breaking it into 4 modules. New ones are Control.Iterate.BaseTypes, Control.Iterate.Exp, and Control.Iterate.BiMap
