+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/Rewards.hs
@@ -102,10 +102,10 @@ import Cardano.Ledger.Slot (epochInfoSize)
 import Cardano.Ledger.Val (invert, (<+>), (<->))
 import Cardano.Prelude (rightToMaybe)
 import Cardano.Slotting.Slot (EpochSize (..))
-import Control.Iterate.SetAlgebra (eval, (◁))
 import Control.Monad (replicateM)
 import Control.Monad.Trans.Reader (asks, runReader)
 import Control.Provenance (ProvM, preservesJust, preservesNothing, runProvM, runWithProvM)
+import Control.SetAlgebra (eval, (◁))
 import Control.State.Transition.Trace (SourceSignalTarget (..), sourceSignalTargets)
 import qualified Data.Compact.VMap as VMap
 import Data.Default.Class (Default (def))
